[Polishing the core] - Renaming ProcessKeystoreGenerator and more code cleanup and Javadoc
