+++ core/src/main/java/org/jboss/intersmash/provision/ProvisionerFactory.java
@@ -17,6 +17,10 @@ package org.jboss.intersmash.provision;
 
 import org.jboss.intersmash.application.Application;
 
+/**
+ * Define the contract for classes that return a suitable {@link Provisioner} for a given {@link Application} class.
+ * @param <T> The type of the {@link Provisioner} that the factory implementation shall return.
+ */
 public interface ProvisionerFactory<T extends Provisioner> {
 
 	/**
