+++ core/src/main/java/org/jboss/intersmash/provision/openshift/FailFastUtils.java
@@ -22,6 +22,10 @@ import cz.xtf.core.openshift.OpenShifts;
 import cz.xtf.core.waiting.failfast.FailFastBuilder;
 import cz.xtf.core.waiting.failfast.FailFastCheck;
 
+/**
+ * Helper class that leverages the XTF library fail-fast APIs in methods that can be used to control the
+ * Intersmash provisioning workflow.
+ */
 public class FailFastUtils {
 	private static String[] failFastEventMessages = new String[] {
 			"Failed to pull image.*",
